+++ src/main/java/faang/school/achievement/messaging/listener/RecommendationEventListener.java
@@ -1,8 +1,8 @@
-package faang.school.achievement.listener;
+package faang.school.achievement.messaging.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.achievement.dto.RecommendationEventDto;
-import faang.school.achievement.handler.EventHandler;
+import faang.school.achievement.messaging.handler.EventHandler;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.connection.Message;
